requirements.txt
torch
torchvision
albumentations
opencv-python
Pillow
wandb
scikit-learn
segmentation_models_pytorch
python-dotenv